+++ CHANGELOG.md
@@ -1,6 +1,25 @@
 # Change Log
 All notable changes to the speechmarkdown-js project will be documented in this file.
 
+## 0.7.0-alpha.0 - (July 6, 2019)
+### Added
+- Support for `audio` tag
+
+## 0.6.0-alpha.0 - (July 6, 2019)
+### Added
+- Support for `voice` and `lang` tags
+
+## 0.5.0-alpha.0 - (July 5, 2019)
+### Fixed
+- Issue #7 - Grammar - multiple modifiers for the same text
+### Added
+- Grammar and formatters for standard:
+    - volume / vol
+    - rate
+    - pitch
+    - sub
+    - ipa
+
 ## 0.4.0-alpha.0 - (June 30, 2019)
 ### Added
 - Update grammar and formatters for standard: